+++ spring-data-jdbc/src/main/java/org/springframework/data/jdbc/core/JdbcAggregateOperations.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017-2021 the original author or authors.
+ * Copyright 2017-2022 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -26,6 +26,7 @@ import org.springframework.lang.Nullable;
  * @author Jens Schauder
  * @author Thomas Lang
  * @author Milan Milanov
+ * @author Chirag Tailor
  */
 public interface JdbcAggregateOperations {
 
@@ -38,6 +39,14 @@ public interface JdbcAggregateOperations {
 	 */
 	<T> T save(T instance);
 
+	/**
+	 * Saves all aggregate instances, including all the members of each aggregate instance.
+	 *
+	 * @param instances the aggregate roots to be saved. Must not be {@code null}.
+	 * @param <T> the type of the aggregate root.
+	 * @return the saved instances.
+	 * @since 3.0
+	 */
 	<T> Iterable<T> saveAll(Iterable<T> instances);
 
 	/**
