+++ spring-data-jdbc/src/test/java/org/springframework/data/jdbc/repository/JdbcRepositoryIntegrationTests.java
@@ -16,6 +16,7 @@
 package org.springframework.data.jdbc.repository;
 
 import static java.util.Arrays.*;
+import static java.util.Collections.*;
 import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.SoftAssertions.*;
 import static org.springframework.test.context.TestExecutionListeners.MergeMode.*;
@@ -50,6 +51,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
+import org.springframework.data.relational.core.mapping.MappedCollection;
 import org.springframework.data.relational.repository.Lock;
 import org.springframework.data.jdbc.repository.query.Modifying;
 import org.springframework.data.jdbc.repository.query.Query;
@@ -63,6 +65,7 @@ import org.springframework.data.relational.core.mapping.event.AfterConvertEvent;
 import org.springframework.data.relational.core.mapping.event.AfterLoadEvent;
 import org.springframework.data.relational.core.sql.LockMode;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.data.repository.core.NamedQueries;
 import org.springframework.data.repository.core.support.PropertiesBasedNamedQueries;
 import org.springframework.data.repository.query.Param;
@@ -80,6 +83,7 @@ import org.springframework.transaction.annotation.Transactional;
  *
  * @author Jens Schauder
  * @author Mark Paluch
+ * @author Chirag Tailor
  */
 @Transactional
 @TestExecutionListeners(value = AssumeFeatureTestExecutionListener.class, mergeMode = MERGE_WITH_DEFAULTS)
@@ -89,6 +93,7 @@ public class JdbcRepositoryIntegrationTests {
 	@Autowired NamedParameterJdbcTemplate template;
 	@Autowired DummyEntityRepository repository;
 	@Autowired MyEventListener eventListener;
+	@Autowired RootRepository rootRepository;
 
 	private static DummyEntity createDummyEntity() {
 
@@ -282,7 +287,7 @@ public class JdbcRepositoryIntegrationTests {
 				.containsExactlyInAnyOrder(entity.getName(), other.getName());
 	}
 
-	@Test
+	@Test // GH-537
 	void insertsOrUpdatesManyEntities() {
 
 		DummyEntity entity = repository.save(createDummyEntity());
@@ -589,6 +594,84 @@ public class JdbcRepositoryIntegrationTests {
 		assertThat(repository.returnInput(null)).isNull();
 	}
 
+	@Test // GH-537
+	void manyInsertsWithNestedEntities() {
+		Root root1 = createRoot("root1");
+		Root root2 = createRoot("root2");
+
+		List<Root> savedRoots = rootRepository.saveAll(asList(root1, root2));
+
+		List<Root> reloadedRoots = rootRepository.findAll();
+		assertThat(reloadedRoots).isEqualTo(savedRoots);
+		assertThat(reloadedRoots).hasSize(2);
+		assertIsEqualToWithNonNullIds(reloadedRoots.get(0), root1);
+		assertIsEqualToWithNonNullIds(reloadedRoots.get(1), root2);
+	}
+
+	@Test // GH-537
+	@EnabledOnFeature(TestDatabaseFeatures.Feature.SUPPORTS_GENERATED_IDS_IN_REFERENCED_ENTITIES)
+	void manyUpdatesWithNestedEntities() {
+		Root root1 = createRoot("root1");
+		Root root2 = createRoot("root2");
+		List<Root> roots = rootRepository.saveAll(asList(root1, root2));
+		Root savedRoot1 = roots.get(0);
+		Root updatedRoot1 = new Root(savedRoot1.id, "updated" + savedRoot1.name,
+				new Intermediate(savedRoot1.intermediate.id, "updated" + savedRoot1.intermediate.name,
+						new Leaf(savedRoot1.intermediate.leaf.id, "updated" + savedRoot1.intermediate.leaf.name), emptyList()),
+				savedRoot1.intermediates);
+		Root savedRoot2 = roots.get(1);
+		Root updatedRoot2 = new Root(savedRoot2.id, "updated" + savedRoot2.name, savedRoot2.intermediate,
+				singletonList(
+						new Intermediate(savedRoot2.intermediates.get(0).id, "updated" + savedRoot2.intermediates.get(0).name, null,
+								singletonList(new Leaf(savedRoot2.intermediates.get(0).leaves.get(0).id,
+										"updated" + savedRoot2.intermediates.get(0).leaves.get(0).name)))));
+
+		List<Root> updatedRoots = rootRepository.saveAll(asList(updatedRoot1, updatedRoot2));
+
+		List<Root> reloadedRoots = rootRepository.findAll();
+		assertThat(reloadedRoots).isEqualTo(updatedRoots);
+		assertThat(reloadedRoots).containsExactly(updatedRoot1, updatedRoot2);
+	}
+
+	@Test // GH-537
+	@EnabledOnFeature(TestDatabaseFeatures.Feature.SUPPORTS_GENERATED_IDS_IN_REFERENCED_ENTITIES)
+	void manyInsertsAndUpdatesWithNestedEntities() {
+		Root root1 = createRoot("root1");
+		Root savedRoot1 = rootRepository.save(root1);
+		Root updatedRoot1 = new Root(savedRoot1.id, "updated" + savedRoot1.name,
+				new Intermediate(savedRoot1.intermediate.id, "updated" + savedRoot1.intermediate.name,
+						new Leaf(savedRoot1.intermediate.leaf.id, "updated" + savedRoot1.intermediate.leaf.name), emptyList()),
+				savedRoot1.intermediates);
+		Root root2 = createRoot("root2");
+		List<Root> savedRoots = rootRepository.saveAll(asList(updatedRoot1, root2));
+
+		List<Root> reloadedRoots = rootRepository.findAll();
+		assertThat(reloadedRoots).isEqualTo(savedRoots);
+		assertThat(reloadedRoots.get(0)).isEqualTo(updatedRoot1);
+		assertIsEqualToWithNonNullIds(reloadedRoots.get(1), root2);
+	}
+
+	private Root createRoot(String namePrefix) {
+		return new Root(null, namePrefix,
+				new Intermediate(null, namePrefix + "Intermediate", new Leaf(null, namePrefix + "Leaf"), emptyList()),
+				singletonList(new Intermediate(null, namePrefix + "QualifiedIntermediate", null,
+						singletonList(new Leaf(null, namePrefix + "QualifiedLeaf")))));
+	}
+
+	private void assertIsEqualToWithNonNullIds(Root reloadedRoot1, Root root1) {
+		assertThat(reloadedRoot1.id).isNotNull();
+		assertThat(reloadedRoot1.name).isEqualTo(root1.name);
+		assertThat(reloadedRoot1.intermediate.id).isNotNull();
+		assertThat(reloadedRoot1.intermediate.name).isEqualTo(root1.intermediate.name);
+		assertThat(reloadedRoot1.intermediates.get(0).id).isNotNull();
+		assertThat(reloadedRoot1.intermediates.get(0).name).isEqualTo(root1.intermediates.get(0).name);
+		assertThat(reloadedRoot1.intermediate.leaf.id).isNotNull();
+		assertThat(reloadedRoot1.intermediate.leaf.name).isEqualTo(root1.intermediate.leaf.name);
+		assertThat(reloadedRoot1.intermediates.get(0).leaves.get(0).id).isNotNull();
+		assertThat(reloadedRoot1.intermediates.get(0).leaves.get(0).name)
+				.isEqualTo(root1.intermediates.get(0).leaves.get(0).name);
+	}
+
 	private Instant createDummyBeforeAndAfterNow() {
 
 		Instant now = Instant.now();
@@ -693,6 +776,11 @@ public class JdbcRepositoryIntegrationTests {
 		}
 
 		@Bean
+		RootRepository rootRepository() {
+			return factory.getRepository(RootRepository.class);
+		}
+
+		@Bean
 		NamedQueries namedQueries() throws IOException {
 
 			PropertiesFactoryBean properties = new PropertiesFactoryBean();
@@ -705,6 +793,30 @@ public class JdbcRepositoryIntegrationTests {
 		MyEventListener eventListener() {
 			return new MyEventListener();
 		}
+	}
+
+	interface RootRepository extends ListCrudRepository<Root, Long> {}
+
+	@Value
+	static class Root {
+		@Id Long id;
+		String name;
+		Intermediate intermediate;
+		@MappedCollection(idColumn = "ROOT_ID", keyColumn = "ROOT_KEY") List<Intermediate> intermediates;
+	}
+
+	@Value
+	static class Intermediate {
+		@Id Long id;
+		String name;
+		Leaf leaf;
+		@MappedCollection(idColumn = "INTERMEDIATE_ID", keyColumn = "INTERMEDIATE_KEY") List<Leaf> leaves;
+	}
+
+	@Value
+	static class Leaf {
+		@Id Long id;
+		String name;
 	}
 
 	static class MyEventListener implements ApplicationListener<AbstractRelationalEvent<?>> {
