+++ spring-data-jdbc/src/test/resources/org.springframework.data.jdbc.repository/JdbcRepositoryIntegrationTests-db2.sql
@@ -1,4 +1,7 @@
 DROP TABLE dummy_entity;
+DROP TABLE ROOT;
+DROP TABLE INTERMEDIATE;
+DROP TABLE LEAF;
 
 CREATE TABLE dummy_entity
 (
@@ -8,4 +11,26 @@ CREATE TABLE dummy_entity
     OFFSET_DATE_TIME TIMESTAMP, -- with time zone is only supported with z/OS
     FLAG             BOOLEAN,
     REF              BIGINT
+);
+
+CREATE TABLE ROOT
+(
+    ID   BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
+    NAME VARCHAR(100)
+);
+CREATE TABLE INTERMEDIATE
+(
+    ID       BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
+    NAME     VARCHAR(100),
+    ROOT     BIGINT,
+    ROOT_ID  BIGINT,
+    ROOT_KEY INTEGER
+);
+CREATE TABLE LEAF
+(
+    ID               BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
+    NAME             VARCHAR(100),
+    INTERMEDIATE     BIGINT,
+    INTERMEDIATE_ID  BIGINT,
+    INTERMEDIATE_KEY INTEGER
 );
