+++ spring-data-jdbc/src/test/resources/org.springframework.data.jdbc.repository/JdbcRepositoryIntegrationTests-mysql.sql
@@ -10,3 +10,26 @@ CREATE TABLE DUMMY_ENTITY
     FLAG             BIT(1),
     REF              BIGINT
 );
+
+CREATE TABLE ROOT
+(
+    ID   BIGINT AUTO_INCREMENT PRIMARY KEY,
+    NAME VARCHAR(100)
+);
+CREATE TABLE INTERMEDIATE
+(
+    ID       BIGINT AUTO_INCREMENT PRIMARY KEY,
+    NAME     VARCHAR(100),
+    ROOT     BIGINT,
+    ROOT_ID  BIGINT,
+    ROOT_KEY INTEGER
+);
+CREATE TABLE LEAF
+(
+    ID               BIGINT AUTO_INCREMENT PRIMARY KEY,
+    NAME             VARCHAR(100),
+    INTERMEDIATE     BIGINT,
+    INTERMEDIATE_ID  BIGINT,
+    INTERMEDIATE_KEY INTEGER
+);
+
