+++ spring-data-jdbc/src/test/resources/org.springframework.data.jdbc.repository/JdbcRepositoryIntegrationTests-oracle.sql
@@ -1,4 +1,7 @@
 DROP TABLE DUMMY_ENTITY CASCADE CONSTRAINTS PURGE;
+DROP TABLE ROOT CASCADE CONSTRAINTS PURGE;
+DROP TABLE INTERMEDIATE CASCADE CONSTRAINTS PURGE;
+DROP TABLE LEAF CASCADE CONSTRAINTS PURGE;
 
 CREATE TABLE DUMMY_ENTITY
 (
@@ -8,4 +11,26 @@ CREATE TABLE DUMMY_ENTITY
     OFFSET_DATE_TIME TIMESTAMP WITH TIME ZONE,
     FLAG             NUMBER(1,0),
     REF              NUMBER
+);
+
+CREATE TABLE ROOT
+(
+    ID   NUMBER GENERATED BY DEFAULT ON NULL AS IDENTITY PRIMARY KEY,
+    NAME VARCHAR2(100)
+);
+CREATE TABLE INTERMEDIATE
+(
+    ID       NUMBER GENERATED BY DEFAULT ON NULL AS IDENTITY PRIMARY KEY,
+    NAME     VARCHAR2(100),
+    ROOT     NUMBER,
+    ROOT_ID  NUMBER,
+    ROOT_KEY NUMBER
+);
+CREATE TABLE LEAF
+(
+    ID               NUMBER GENERATED BY DEFAULT ON NULL AS IDENTITY PRIMARY KEY,
+    NAME             VARCHAR2(100),
+    INTERMEDIATE     NUMBER,
+    INTERMEDIATE_ID  NUMBER,
+    INTERMEDIATE_KEY NUMBER
 );
