--- spring-data-jdbc/src/test/resources/org.springframework.data.jdbc.repository/JdbcRepositorySaveAllHsqlIntegrationTests-hsql.sql
@@ -1,21 +1,0 @@
-CREATE TABLE ROOT
-(
-    ID   BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
-    NAME VARCHAR(100)
-);
-CREATE TABLE INTERMEDIATE
-(
-    ID       BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
-    NAME     VARCHAR(100),
-    ROOT     BIGINT,
-    ROOT_ID  BIGINT,
-    ROOT_KEY INTEGER
-);
-CREATE TABLE LEAF
-(
-    ID               BIGINT GENERATED BY DEFAULT AS IDENTITY ( START WITH 1 ) PRIMARY KEY,
-    NAME             VARCHAR(100),
-    INTERMEDIATE     BIGINT,
-    INTERMEDIATE_ID  BIGINT,
-    INTERMEDIATE_KEY INTEGER
-);
