+++ spring-data-relational/src/main/java/org/springframework/data/relational/core/conversion/MutableAggregateChange.java
@@ -74,6 +74,14 @@ public interface MutableAggregateChange<T> extends AggregateChange<T> {
 		return forDelete((Class<T>) ClassUtils.getUserClass(entity));
 	}
 
+	/**
+	 * Factory method to create a {@link MergedAggregateChange} for saving entities.
+	 *
+	 * @param entityClass aggregate root type.
+	 * @param <T> entity type.
+	 * @return the {@link MergedAggregateChange} for saving root entities.
+	 * @since 3.0
+	 */
 	static <T> MergedAggregateChange<T, AggregateChangeWithRoot<T>> mergedSave(Class<T> entityClass) {
 
 		Assert.notNull(entityClass, "Entity class must not be null");
