+++ spring-data-relational/src/test/java/org/springframework/data/relational/core/conversion/SaveMergedAggregateChangeTest.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2020-2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.springframework.data.relational.core.conversion;
 
 import static java.util.Collections.*;
@@ -14,6 +29,11 @@ import org.springframework.data.relational.core.mapping.RelationalMappingContext
 
 import lombok.Value;
 
+/**
+ * Unit tests for {@link SaveMergedAggregateChange}.
+ *
+ * @author Chirag Tailor
+ */
 class SaveMergedAggregateChangeTest {
 
 	RelationalMappingContext context = new RelationalMappingContext();
